+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `JsonSigner` enum variants to `HttpUrl` and `DidUrl`.
+
 ## [0.4.0] - 2024-11-07
 
 ### Added